+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.6.0-beta.0] - Unreleased
 
+### Added
+- Added `esp:get_default_mac/0` for retrieving the default MAC address on ESP32.
+
 ## [0.6.0-alpha.2] - 2023-12-10
 
 ### Fixed